--- src/SharedLib/SharedLib.sln
@@ -1,25 +1,0 @@
-﻿
-Microsoft Visual Studio Solution File, Format Version 12.00
-# Visual Studio Version 17
-VisualStudioVersion = 17.4.33122.133
-MinimumVisualStudioVersion = 10.0.40219.1
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "SharedLib", "SharedLib\SharedLib.csproj", "{53802764-C19A-4696-BFB9-DBECEDECCEAB}"
-EndProject
-Global
-	GlobalSection(SolutionConfigurationPlatforms) = preSolution
-		Debug|Any CPU = Debug|Any CPU
-		Release|Any CPU = Release|Any CPU
-	EndGlobalSection
-	GlobalSection(ProjectConfigurationPlatforms) = postSolution
-		{53802764-C19A-4696-BFB9-DBECEDECCEAB}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{53802764-C19A-4696-BFB9-DBECEDECCEAB}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{53802764-C19A-4696-BFB9-DBECEDECCEAB}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{53802764-C19A-4696-BFB9-DBECEDECCEAB}.Release|Any CPU.Build.0 = Release|Any CPU
-	EndGlobalSection
-	GlobalSection(SolutionProperties) = preSolution
-		HideSolutionNode = FALSE
-	EndGlobalSection
-	GlobalSection(ExtensibilityGlobals) = postSolution
-		SolutionGuid = {9E6EC375-EAE7-402C-9943-FA4B420AEE8A}
-	EndGlobalSection
-EndGlobal
